Change: Add warning about relative dirs for bind mounts
The documentation was not entirely clear about the need to create directories for bind mounts for the database and application containers, so I've added some context hear for the avoidance of doubt.

I've also added a warning in each case to ensure that fully-qualified directory names are passed.  If you pass a multi-level relative path, you'll get an error message, but if you specify only a single word for a directory, there's an amusing edge case with Docker Desktop, where this gets used as a name, and the mount is managed inside the hidden docker desktop container, which can cause confusing side-effects.